Some minor bugfixes.  Also, I'm uploading a preliminary version of my in place category changing option.  Rather than shuffling everything to the bottom (which breaks when categories are included inside of tags such as <noinclude>), this should fix everything in place.  The only problem is that it doesn't work yet at all.  So don't use the -inplace option just yet ... it won't work.  All current uses of category.py should be unaffected though.
Please look at my attempt to get inplace replacement working and fix it as
necessary.  This will be a very useful functionality to have - I just haven't
figured it out yet.  All of the framework to get it working is now in place,
you just need to fix up that one function in wikipedia.py that I've added.

https://mediawiki.org/wiki/Special:Code/pywikipedia/2672