Another important bugfix.
There was a whitespace problem in the XML parser. A newline and some spaces were 
added to the revision ID number, and since a change to the MediaWiki Flagged 
Revisions extension, this whitespace is a problem. Reviewed articles on dewiki had 
to be reviewed again after a bot made a mistake.

If you have a bot that makes changes on dewiki, please update immediately, as many 
users are complaining about the bug.


https://mediawiki.org/wiki/Special:Code/pywikipedia/5387