Added the possibility to work directly on the compressed .bz2 file. It is a bit slower, of course.
The rationale behind this is that it is becoming difficult to unpack the XML dumps. For example, the current pages-articles.xml for 
the German Wikipedia is over 4 GB big, so you can no longer unpack it to a FAT32 partition. And even if you have another partition, 
the dump will use very much disk space.


https://mediawiki.org/wiki/Special:Code/pywikipedia/5775