Add support for namespace-filtered page generators.
Before, generators had to get data for /all/ namespaces, and these results
were filtered. This commit adds support for communicating wanted namespaces
with generators.

Instead of creating a generator by passing a list of namespaces, a /function/
that returns a list is passed. This function is only called once the generator
starts generating Page objects, which is /after/ the GeneratorFactory has
completed parsing all command line arguments. Then, the getNamespaces function
is called, and the correct namespaces are used.


https://mediawiki.org/wiki/Special:Code/pywikipedia/11383