hostname and paths should be str in generated family files
All parameters passed to request should be bytestrings - str. If any
of them is accidentally unicode, this will try to convert all bytestrings
to unicode via str.decode(sys.defaultencoding). This is problematic,
because non-ascii data will then throw an UnicodeDecodeError as shown in [1].

The url will be converted to str on the postData level, but it is also good
to prevent it from being unicode in the first place.

See also: I1b66a31c8752dee1f950d425d532479f1b671926

[1] http://lists.wikimedia.org/pipermail/pywikipedia-l/2013-August/008218.html

Change-Id: I5eb13d8853b6ad08b48f40b5ad881616a2f4fd2a