Enabling usage of Disambiguator
Per help page [1] and some examples [2] [3] when someone runs isDisambig method this code checks if the Disambiguator is enabled,
By checking list of extensions in siteinfo, if the extension is not enabaled (like [4]) returns the old way
but if exists (like [5]) uses API to check being disambiguation. For more info see the examples I showed in [2] and [3]

It's prttey important, and please review as soon as possible because if the extension is enabled, the old way doens't work now.
See the bug.
[1]: https://www.mediawiki.org/wiki/Extension:Disambiguator
[2]: https://en.wikipedia.org/w/api.php?action=query&titles=Aa%20River&prop=pageprops&ppprop=disambiguation&format=jsonfm
[3]: https://en.wikipedia.org/w/api.php?action=query&titles=Tehran&prop=pageprops&ppprop=disambiguation&format=jsonfm
[4]: http://tools.wmflabs.org/wikitest-rtl/w/index.php?title=Special:DisambiguationPages
[5]: https://en.wikipedia.org/wiki/Special:DisambiguationPages

Bug: 54480
Change-Id: Ie3c46671d888cf917640fc732a28cc827028d0de