(bugfix) imagetransfer.py: remove broken debug
In the conversion from lib_images.py to imagetransfer.py,
the following line was dropped:
    imageTitle = sourceImagePage.title().encode('utf-8')

https://www.mediawiki.org/wiki/Special:Code/pywikipedia/2066
sha 9f6e1eb090ee6ee355cf65984cd3adcf851c8470

As a result, the logging of the imageTitle has referenced an
undefined variable since 2005.  This debug log entry is disabled
in the code. There are also no bugs reported, so it is unlikely to
be needed.

Change-Id: Ie044fba2e32e1f34e9d96eb4baa45bd52fc04e8b